+++ src/wai/json/object/property/_BoolProperty.py
@@ -1,6 +1,7 @@
 from typing import Optional
 
 from ...schema import BOOL_SCHEMA
+from .._typing import PropertyValueType, Absent
 from ._RawProperty import RawProperty
 
 
@@ -11,9 +12,11 @@ class BoolProperty(RawProperty):
     def __init__(self,
                  name: Optional[str] = None,
                  *,
-                 optional: bool = False):
+                 optional: bool = False,
+                 default: PropertyValueType = Absent):
         super().__init__(
             name,
             schema=BOOL_SCHEMA,
-            optional=optional
+            optional=optional,
+            default=default
         )