+++ src/wai/json/object/property/_ProxyProperty.py
@@ -6,7 +6,7 @@ from wai.common.abc import is_abstract_class
 from ...error import JSONPropertyError
 from ...schema import JSONSchema
 from ...serialise import JSONValidatedBiserialisable
-from .._typing import PropertyValueType
+from .._typing import PropertyValueType, Absent
 from ._Property import Property
 
 
@@ -19,18 +19,20 @@ class ProxyProperty(Property, ABC):
                  name: Optional[str] = None,
                  *,
                  proxy: Type[JSONValidatedBiserialisable] = JSONValidatedBiserialisable,
-                 optional: bool = False):
-        super().__init__(
-            name,
-            optional=optional
-        )
-
+                 optional: bool = False,
+                 default: PropertyValueType = Absent):
         # Type argument must be concrete
         if is_abstract_class(proxy):
             raise JSONPropertyError(f"Proxy type argument must be a concrete class, got {proxy.__name__}")
 
         # The type of values this property takes
         self._type: Type[JSONValidatedBiserialisable] = proxy
+
+        super().__init__(
+            name,
+            optional=optional,
+            default=default
+        )
 
     @property
     def proxy_type(self) -> Type[JSONValidatedBiserialisable]: